+++ themes/processing/src/.babelrc
@@ -1,5 +1,3 @@
 {
-  "presets": [
-    ["env"]
-  ]
+  "presets": [ "@babel/preset-env" ]
 }
