--- themes/processing/src/js/lib/math.js
@@ -1,6 +1,0 @@
-// lib/math.js
-export function sum(x, y) {
-  return x + y;
-}
-
-export var pi = 3.141593;
